Cleanup and improvement.
